correctly set hours in eastern
